requirements.txt
requests>=2.26.0
aiohttp>=3.9.1
httpx>=0.25.0
fastapi>=0.68.0
uvicorn>=0.15.0

prefect>=2.0.0

pydantic>=2.0.0
click>=8.0.0

thefuzz[speedup]>=0.19.0
python-Levenshtein>=0.21.0
textstat>=0.7.3

python-dotenv>=0.19.0
backoff>=2.2.1
asyncio>=3.4.3

pandas>=2.0.0
PyYAML>=6.0.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pyright>=1.1.0

nltk>=3.8.1
ftfy>=6.1.1
ruff>=0.1.0